Add Control builder
