+++ .gitignore
@@ -1,7 +1,10 @@
+.env
+node_modules/
+
 assets/css/*
 assets/js/*
-
-node_modules/
+library/backend/assets/js/greenlet-controls.js
+library/backend/assets/fonts/google-fonts.json
 
 **/cypress.env.json
 **/cypress/videos
@@ -10,6 +13,3 @@ node_modules/
 **/cypress/fixtures
 
 **/lighthouse/reports
-
-.env
-library/backend/assets/fonts/google-fonts.json
